+++ src/models/getprojectsvaluecurrentvalue.ts
@@ -413,6 +413,7 @@ export type ResponseBodyPermissions = {
   marketplaceBillingData?: Array<ACLAction> | undefined;
   marketplaceExperimentationEdgeConfigData?: Array<ACLAction> | undefined;
   marketplaceExperimentationItem?: Array<ACLAction> | undefined;
+  marketplaceFlexCommit?: Array<ACLAction> | undefined;
   marketplaceInstallationMember?: Array<ACLAction> | undefined;
   marketplaceInvoice?: Array<ACLAction> | undefined;
   marketplaceSettings?: Array<ACLAction> | undefined;
@@ -2230,6 +2231,7 @@ export const ResponseBodyPermissions$inboundSchema: z.ZodType<
   marketplaceExperimentationItem: types.optional(
     z.array(ACLAction$inboundSchema),
   ),
+  marketplaceFlexCommit: types.optional(z.array(ACLAction$inboundSchema)),
   marketplaceInstallationMember: types.optional(
     z.array(ACLAction$inboundSchema),
   ),