+++ src/models/getprojectvalue.ts
@@ -321,6 +321,7 @@ export type GetProjectPermissions = {
   marketplaceBillingData?: Array<ACLAction> | undefined;
   marketplaceExperimentationEdgeConfigData?: Array<ACLAction> | undefined;
   marketplaceExperimentationItem?: Array<ACLAction> | undefined;
+  marketplaceFlexCommit?: Array<ACLAction> | undefined;
   marketplaceInstallationMember?: Array<ACLAction> | undefined;
   marketplaceInvoice?: Array<ACLAction> | undefined;
   marketplaceSettings?: Array<ACLAction> | undefined;
@@ -1892,6 +1893,7 @@ export const GetProjectPermissions$inboundSchema: z.ZodType<
   marketplaceExperimentationItem: types.optional(
     z.array(ACLAction$inboundSchema),
   ),
+  marketplaceFlexCommit: types.optional(z.array(ACLAction$inboundSchema)),
   marketplaceInstallationMember: types.optional(
     z.array(ACLAction$inboundSchema),
   ),