+++ src/models/updateprojectlogheaders.ts
@@ -810,6 +810,7 @@ export type UpdateProjectProjectsBuildMachineSelection = ClosedEnum<
 >;
 
 export const UpdateProjectProjectsBuildMachineElasticReason = {
+  BasicFloor: "basic-floor",
   BuildTimeoutFailure: "build-timeout-failure",
   EnospcFailure: "enospc-failure",
   EnterpriseFloor: "enterprise-floor",
@@ -949,6 +950,7 @@ export type UpdateProjectProjectsResponseBuildMachineSelection = ClosedEnum<
 >;
 
 export const UpdateProjectProjectsResponseBuildMachineElasticReason = {
+  BasicFloor: "basic-floor",
   BuildTimeoutFailure: "build-timeout-failure",
   EnospcFailure: "enospc-failure",
   EnterpriseFloor: "enterprise-floor",
@@ -1314,6 +1316,7 @@ export type UpdateProjectPermissions = {
   marketplaceBillingData?: Array<ACLAction> | undefined;
   marketplaceExperimentationEdgeConfigData?: Array<ACLAction> | undefined;
   marketplaceExperimentationItem?: Array<ACLAction> | undefined;
+  marketplaceFlexCommit?: Array<ACLAction> | undefined;
   marketplaceInstallationMember?: Array<ACLAction> | undefined;
   marketplaceInvoice?: Array<ACLAction> | undefined;
   marketplaceSettings?: Array<ACLAction> | undefined;
@@ -3786,6 +3789,7 @@ export const UpdateProjectPermissions$inboundSchema: z.ZodType<
   marketplaceExperimentationItem: types.optional(
     z.array(ACLAction$inboundSchema),
   ),
+  marketplaceFlexCommit: types.optional(z.array(ACLAction$inboundSchema)),
   marketplaceInstallationMember: types.optional(
     z.array(ACLAction$inboundSchema),
   ),