+++ RELEASES.md
@@ -1609,3 +1609,13 @@ Based on:
 - [typescript v1.28.14] .
 ### Releases
 - [NPM v1.28.14] https://www.npmjs.com/package/@vercel/sdk/v/1.28.14 - .
+
+## 2026-08-01 00:23:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.28.15] .
+### Releases
+- [NPM v1.28.15] https://www.npmjs.com/package/@vercel/sdk/v/1.28.15 - .